[util.amplitude] ignore number of threads when estimating time
If a single integral takes much longer that the other integrals that are
running in parallel, then it will eventually be awaited before new
integrations are launched. This behavior is on purpose to circumvent
an inefficient initialization where only few samples are needed and
therefore multiple integrals should run concurrently.