Add form_memory_use and form_threads to *_package
This allows to predictably bound the memory usage of TFORM, while
allowing for it to use as much resources as possible.

Note that with the default form_workspace="500M" and form_threads=2,
the bound on memory usage is ~6GB. FORM itself defaults to
WorkSpace being '40M', which bounds the memory at 1.1G.

The code for formset.py comes from [1], commit aa52c0b.

[1] https://github.com/tueda/formset