[codegen] Rename `mv.py` into `move` to unbreak Python 3
Previously setup.py would pre-compile mv.py into mv.pyc; because
mv.py is under code_writer/templates, mv.pyc would be considered
a template by parse_template_tree(), so it would try to read
it and apply replacements. This fails under Python 3, because
mv.pyc is not valid UTF-8, so one can't read it as text.

The solution is to rename `mv.py` into `move`, so that setup.py
wouldn't try to precompile it.