[sum_package] simplify Makefiles, fix GiNaC linking again
The previous fix worked on Linux, but not on MacOS. This time try
to avoid the problem by using static libraries on intermediate
stages.

Also add a static library for the sum_package integrals, and
remove some of the nested Makefiles (it is always best to have
fewer of these).