[code_writer] make integrate_<name> a useful CLI program
Make ./integrate_<name> take the integral parameters as CLI
arguments, so one can just compile it directly without the need
to edit the source and insert the parameters there. It still has
no provision to choose e.g. the requested precision though, so
some editing will be needed to make it practical.