[make_package] flatten Makefiles, allow for more parallel compilation
Previously there was a choke point during compilation: any
compilation of generated code could start only when *all* of
the code was generated. This put a limit on how parallel the
compilation could be.

The reason was that write_integrand.frm produces different number
of files per sector depending on pole order, and the list of
these files was not available in the Makefile.

The solution is to:

- introduce codegen/sector<i>.d which list precisely which files
  will be generated for a sector -- to be included by the main
  Makefile;

- combine all Makefiles into one, so that the full dependency
  graph is available to make; this also makes the Makefile easier
  to understand for a human.

The end result is that the compilation time of examples/box1L
improves from 11.5s at an average 500% CPU utilization to 7.5s at
an average 850% CPU utilization. As expected, more parallelization
is possible when the full dependency graph is available.