[qmc] allow 0 cputhreads, default to len(os.sched_getaffinity(0))
In a cluster setup only a fraction of all CPUs is normally
available to a program; in this case it is best to default Qmc
`cputhreads` to the number of allocated CPUs, rather than the
total count. On Linux it is often possible to get this number
via sched_getaffinity(), so lets use that as the default.

Note that the default is for the Python interface only: for the
C++ interface to do the same we would need to first detect if
the platform has sched_getaffinity() at all, which would need a
separate configuration step.

Additionally, allow `cputhreads` to be set to 0, disabling CPU
evaluation (in favor of GPU).