[sum_package] allow parallel generation of the terms in a sum
Add `processes` to `sum_package()`, so that the choice is between
generating each term in a single process, but concurrently with
other terms (by passing `processes` to `sum_package()`) and
running all terms sequentially, but each sector within it in
parallel (by passing `processes` to each `MakePackage()` in a
sum). The more terms and the fewer sectors they have, the faster
it is to do the former.