[make_package] rewrite how sources are generated, improve the integrands
Previously the temporary variables FORM generated during
optimization were put into an array, and therefore had to have
the same type; because of this variables that could have been
real were forced to be complex. To get rid of this problem:
transform the array into just local variables, and rewrite the
code such that each variable is only assigned once (previously
each element of the array would be assigned to multiple times).
This allows each variable to have the correct type.

Because this change requires a transformation pass over the
generated code, the way write_integrand.frm generates the code
had to be rewritten: instead of producing .cpp and .hpp files
directly, write_integrand.frm now produces one .info file per
sector, listing the optimized function bodies (along with data
defining the sector), and a separate piece of Python then reads
this file, transforms the code, and writes down .cpp and .hpp
files. This way we do all the string manipulation in Python
rather than FORM.