[deval] load the coefficients faster
Because we rationalize the variable values during the coefficient
evaluations, the coefficient values are giant rationals. It
turns out that often it takes more time to print and parse them
than to actually do the series expansion. The solution is to let
ginsh evaluate the final result, and only print the floating
point values.