[contrib] force GMP to build in generic configuration
By default GMP picks up the processor used during the build,
and optimizes the library for it -- at the expense of outright
crasing on other processors. To disable this behavior, one should
overwrite the processor guessing script 'config.guess' with the
generic processor guessing script 'configfsf.guess' (which is
also shipped with GMP).