[deval] use disteval.py as a library in DistevalLibrary
Previously DistevalLibrary would use the disteval CLI, i.e.
start a whole new Python process for integration. This has an
overhead if multiple points are to be evaluated. To overcome
some of this overhead, disteval.py is now restructured so that
it can be used as a Python library, and DistevalLibrary does
just that.

The added benefit is that the workers are retained between the
evaluation calls, so if multiple evaluations are requested from
the same DistevalLibrary object, the workers will not need to
be re-launched, removing the associated overhead.