[deval] make DistevalLibrary(verbose=X) sticky
DistevalLibrary prints the log in two different places: during
the initialization, and during integration. Both parts have their
own "verbose" flags. With this change the flag used for the
initialization is retained and reused during the integration, if
no other "verbose" value was passed. This way, one only needs
to pass "verbose" once.