[deval] make cudaworker memory fixed at 128MB
Previously it would increase the buffer size for bigger lattices;
now it splits big lattice calculation into batches, so that
any lattice size will use the same memory.

Some quick benchmarks show no performance regression. In fact,
even decreasing the buffer size from 128MB down to 1MB does not
affect the performance. There is a rounding-level change in the
results because the order of summation has changed. This is as
expected.