[deval] fix the reporting of final integration errors
The calculation of the variance of the sum values was wrong in
situations when either the coefficients or the prefactors have
non-zero imaginary parts.

The correct calculation is:

    Var(Coef * Value) =
        { Re(Coef)^2 * Var(Re(Value)) + Im(Coef)^2 * Var(Im(Value)) } +
        { Im(Coef)^2 * Var(Re(Value)) + Re(Coef)^2 * Var(Im(Value)) } i

Previously it was assumed to be:

    Var(Coef * Value) =/=
        Abs(Coef)^2 ( Var(Re(Value)) + i Var(Im(Value)) )

The problem only affects the final reporting of the variance. The
stopping criterion used in the intermediate stages of integration
is instead based on the "absolute variance":

    AbsVar(Coef * Value) =
        Var(Re(Coef * Value)) + Var(Im(Coef * Value)) =
        Abs(Coef)^2 ( Var(Re(Value)) + Var(Im(Value)) )

This part was correct, and remains unchanged.