[deval] allow users to pass rational numbers to disteval
In addition to specifying the variables on the command line as

    python3 -m pySecDec.disteval /path/to/integral.json var=1.234

we will now support

    python3 -m pySecDec.disteval /path/to/integral.json var=1234/1000

In this case the "1234/1000" string is passed directly to GiNaC
when the coefficients are evaluated, so that no precision loss
occurs at any stage.

In the Python interface instead of

    result = integral_library({"var": 1.234})

we will now additionally support

    result = integral_library({"var": "1.234"})

and

    result = integral_library({"var": sympy.sympify("1234/1000")})

Both of these forms are preferred to passing floats, because
the coefficients will be evaluated by converting these values
to str() and passing it to GiNaC: floats may introduce rounding
errors in this process, while strings and sympy expressions will
not.