[deval] fix the order of epsabs/epsrel specification
The previous commit intoduced a mismatch between the order
of "epsrel" array and the order of the amplitudes (i.e. the
"amp_val" array and the like), which manifested in epsrel being
applied to the wrong amplitudes if specified per-amplitude. This
commit fixes the mismatch, while keeping the the printed order
of "epsrel" and "epsabs" stable.