[disteval] fix RandomScheduler.drain()
If the *final* received worker integration result is a `NaN`
and needs to be rescheduled, RandomScheduler.drain() would end
without waiting for the rescheduled integration, because it would
see that there were no pending jobs at one point, and would not
re-check if there are any right now.

This is triggered rarely, and would result in the mean and the
variance calculated incorrectly (likely at an intermediate
stage).

Reported-By: Anton Olsson